@@ -6,7 +6,7 @@
 //
 
 // MARK: - Photos
-public struct Photos: Codable {
+public struct Photos: Codable, Sendable {
     public let photos: [Photo]
 }
 
@@ -15,8 +15,17 @@ public struct Photo: Codable, Identifiable, Equatable, Hashable, Sendable {
     public let id, sol: Int
     public let camera: Camera
     public let imgSrc: String
-    let earthDate: String
-    let rover: Rover
+    public let earthDate: String
+    public let rover: Rover
+
+    public init(id: Int, sol: Int, camera: Camera, imgSrc: String, earthDate: String, rover: Rover) {
+        self.id = id
+        self.sol = sol
+        self.camera = camera
+        self.imgSrc = imgSrc
+        self.earthDate = earthDate
+        self.rover = rover
+    }
 
     enum CodingKeys: String, CodingKey {
         case id, sol, camera
@@ -25,25 +34,3 @@ public struct Photo: Codable, Identifiable, Equatable, Hashable, Sendable {
         case rover
     }
 }
-
-
-//
-//{
-//"id": 102693,
-//"sol": 1000,
-//"camera": {
-//"id": 20,
-//"name": "FHAZ",
-//"rover_id": 5,
-//"full_name": "Front Hazard Avoidance Camera"
-//},
-//"img_src": "http://mars.jpl.nasa.gov/msl-raw-images/proj/msl/redops/ods/surface/sol/01000/opgs/edr/fcam/FLB_486265257EDR_F0481570FHAZ00323M_.JPG",
-//"earth_date": "2015-05-30",
-//"rover": {
-//"id": 5,
-//"name": "Curiosity",
-//"landing_date": "2012-08-06",
-//"launch_date": "2011-11-26",
-//"status": "active"
-//}
-//},
