@@ -9,5 +9,5 @@ import Foundation
 
 public enum APIConfiguration {
     static let apiKey = "hauCFuire0xmOm1JVntGzUITXVWTLVK9mIztmIaF"//"YOUR_API_KEY_HERE"
-    public static let baseUrl = "https://api.nasa.gov/mars-photos/api/v1/rovers"
+    public static let baseUrl = "https://api.nasa.gov/mars-photos/api/v1/"
 }
