@@ -6,15 +6,17 @@
 //
 
 import Foundation
+import SimplySave
 
 public protocol NasaMissionPersitentDataServicing<Value>: Actor {
-    associatedtype Value: Codable & Identifiable
+    associatedtype Value: Codable & Identifiable & Sendable
 
     func add(with newElement: Value)
     func remove(with element: Value)
     func getCurrentContent() -> [Value]
-    func load(from fileNamed: String) throws
-    func save(with fileName: String) throws
+    func load() async throws -> [Value]
+    func persistData() async throws
+    func clear() async throws
 }
 
 public enum PersistenceError: Swift.Error {
@@ -25,12 +27,14 @@ public enum PersistenceError: Swift.Error {
     case readingDataFailed
 }
 
-public actor SOMPersistentDataService<Value: Codable & Identifiable>: NasaMissionPersitentDataServicing {
-    private var currentData = [Value]()
-    private let fileManager: FileManager
+public actor SOMPersistentDataService<Value: Codable & Sendable & Identifiable>: NasaMissionPersitentDataServicing {
 
-    init(fileManager: FileManager = .default) {
-        self.fileManager = fileManager
+    private var currentData = [Value]()
+    let storageManager: SimpleSaving
+    private  let containerName = "missions.json"
+
+    public init(storageManager: SimpleSaving = SimplySaveClient()) {
+        self.storageManager = storageManager
     }
 }
 
@@ -44,50 +48,22 @@ public extension SOMPersistentDataService {
     }
 
     func remove(with element: Value) {
-        currentData.removeAll { $0.id ==  element.id }
+        currentData.removeAll { $0.id == element.id }
     }
 
-    func save(with fileName: String) throws {
-        guard let url = makeURL(forFileNamed: fileName) else {
-            throw PersistenceError.invalidDirectory
-        }
-
-        if fileManager.fileExists(atPath: url.absoluteString) {
-            throw PersistenceError.fileAlreadyExist
-        }
-
-        do {
-            let data = try JSONEncoder().encode(currentData)
-            try data.write(to: url)
-        } catch {
-            debugPrint(error)
-            throw PersistenceError.writtingFailed
-        }
+    func persistData() async throws {
+        try await clear()
+        try await storageManager.save(currentData, as: containerName, in: .documents)
     }
 
-    func load(from fileNamed: String) throws {
-           guard let url = makeURL(forFileNamed: fileNamed) else {
-               throw PersistenceError.invalidDirectory
-           }
-           guard fileManager.fileExists(atPath: url.absoluteString) else {
-               throw PersistenceError.fileDoesNotExist
-           }
-           do {
-               let data = try Data(contentsOf: url)
-               let entries = try JSONDecoder().decode([Value].self, from: data)
-               currentData.append(contentsOf: entries)
-           } catch {
-               debugPrint(error)
-               throw PersistenceError.readingDataFailed
-           }
-       }
-}
+    func load() async throws -> [Value] {
+        let values: [Value] = try await storageManager.fetch(from: containerName, in: .documents)
 
-private extension SOMPersistentDataService {
-    func makeURL(forFileNamed fileName: String) -> URL? {
-        guard let url = fileManager.urls(for: .documentDirectory, in: .userDomainMask).first else {
-            return nil
-        }
-        return url.appendingPathComponent(fileName)
+        currentData.append(contentsOf: values)
+        return currentData
+    }
+
+    func clear() async throws {
+        try await storageManager.remove(containerName, from: .documents)
     }
 }
